CustomErrorType

+++ src/main/java/br/com/ecommerce/controllers/CartController.java
@@ -1,6 +1,8 @@
 package br.com.ecommerce.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -12,6 +14,7 @@ import br.com.ecommerce.models.Item;
 import br.com.ecommerce.models.Product;
 import br.com.ecommerce.services.CartService;
 import br.com.ecommerce.services.ProductService;
+import br.com.ecommerce.util.CustomErrorType;
 
 @RestController
 @RequestMapping("/cart")
@@ -29,40 +32,67 @@ public class CartController {
     }
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
-	public Cart findOne(@PathVariable("id") long id) {
-		return cartService.findById(id);
+	public ResponseEntity<?> findOne(@PathVariable("id") long id) {
+		 Cart currentCart = cartService.findById(id);
+		 
+		 if (currentCart == null) {
+	            return new ResponseEntity(new CustomErrorType("Cart with id " + id 
+	                    + " not found"), HttpStatus.NOT_FOUND);
+	     }
+		 return new ResponseEntity<Cart>(currentCart, HttpStatus.OK);
 	}
 	
 	@RequestMapping(value = "/{idCart}/{idProduto}/{quantity}", method = RequestMethod.GET)
-	public Cart addProductCart(@PathVariable("idCart") long idCart, 
+	public ResponseEntity<?> addProductCart(@PathVariable("idCart") long idCart, 
 			@PathVariable("idProduto") long idProduto, 
 			@PathVariable("quantity") int quantity) {
 		
 		Cart cart = cartService.findById(idCart);
 		Product product = productService.findById(idProduto);
 		
-		return cartService.addProductCart(cart, product, quantity);
+		 if (cart == null) 
+	         return new ResponseEntity(new CustomErrorType("Cart not found"), HttpStatus.NOT_FOUND);
+		 if (product == null)
+			 return new ResponseEntity(new CustomErrorType("Product not found"), HttpStatus.NOT_FOUND);
+		
+		Cart currentCart = cartService.addProductCart(cart, product, quantity);
+		
+		return new ResponseEntity<Cart>(currentCart, HttpStatus.CREATED);
 	}
 	
 	@RequestMapping(value = "/{idCart}/{idProduto}/{quantity}", method = RequestMethod.PUT)
-	public Cart updateQuantityProductCart(@PathVariable("idCart") long idCart, 
+	public ResponseEntity<?> updateQuantityProductCart(@PathVariable("idCart") long idCart, 
 			@PathVariable("idProduto") long idProduto, 
 			@PathVariable("quantity") int quantity) {
 		
 		Cart cart = cartService.findById(idCart);
 		Product product = productService.findById(idProduto);
 		
-		return cartService.updateQuantityProductCart(cart, product, quantity);
+		 if (cart == null) 
+	         return new ResponseEntity(new CustomErrorType("Cart not found"), HttpStatus.NOT_FOUND);
+		 if (product == null)
+			 return new ResponseEntity(new CustomErrorType("Product not found"), HttpStatus.NOT_FOUND);
+		
+		 Cart currentCart = cartService.updateQuantityProductCart(cart, product, quantity);
+		 
+		return new ResponseEntity<Cart>(currentCart, HttpStatus.OK);
 	}
 	
 	@RequestMapping(value = "/{idCart}/{idProduto}", method = RequestMethod.DELETE)
-	public Cart removeProductCart(@PathVariable("idCart") long idCart, 
+	public ResponseEntity<?> removeProductCart(@PathVariable("idCart") long idCart, 
 			@PathVariable("idProduto") long idProduto) {
 		
 		Cart cart = cartService.findById(idCart);
 		Product product = productService.findById(idProduto);
 		
-		return cartService.removeProductCart(cart, product);
+		 if (cart == null) 
+	         return new ResponseEntity(new CustomErrorType("Cart not found"), HttpStatus.NOT_FOUND);
+		 if (product == null)
+			 return new ResponseEntity(new CustomErrorType("Product not found"), HttpStatus.NOT_FOUND);
+		
+		 cartService.removeProductCart(cart, product);
+		 
+		return new ResponseEntity<Cart>(HttpStatus.NO_CONTENT);
 	}
 	
 }
